requirements.txt
flask==2.0.1
scikit-learn==1.0.2
pandas==1.3.3
numpy==1.21.2
jieba==0.42.1
joblib==1.1.0
requests==2.31.0
python-dotenv==1.0.0 